\name{island}
\docType{data}
\alias{island}
\title{rmetasim landscape containing an island migration structure}
\description{
This dataset consists of 10 populations arraged in an island
configuration.  It can be used to prototype functions that operate on
landscapes}
\usage{data(stepping)}
\format{A list in rmetasim landscape format}
\source{Fake data}
\keyword{datasets}
