\name{new.example.landscape}
\alias{new.example.landscape}
\title{Create a Default Landscape}

\description{
  Create a Rmetasim landscape with all default parameters.}
\usage{
  rland <- new.example.landscape()
}
\arguments{
  None
}
\examples{
  ## Only usage
  new.example.landscape()}
\keyword{misc}
