\name{new.landscape.empty}
\alias{new.landscape.empty}
\title{Create a Skeletal Landscape}

\description{
  Create a skeletal Rmetasim landscape ready to be configured}
\usage{
  rland <- new.landscape.empty()
}
\arguments{
  None
}
\examples{
  ## Only usage
  new.landscape.empty()}
\keyword{misc}
