\name{popstruct}
\alias{popstruct}
\title{Calculates Weir and Cokerham's theta}

\description{
Takes data in the format returned by \code{landscape.to.rtheta()} and
returns estimates of Fst in the form of theta }

\usage{
  thetamat <- popstruct(genin)
}
\arguments{
  \item{genin}{a dataframe or matrix with columns corresponding to
  population, class, individual, locus, allelic position, and allele
  ID.  The matrix is most easily generated by using the function
  \code{landscape.to.rtheta()}}
}

\value{
A matrix with two columns.  One gives estimates of theta and the other
estimates of F-hat.  Rows correspond to the different loci.
}
\details{
Calculates theta using a nested analysis of variance approach first
outlined by Weir and Cockerham in 1984}

\seealso{
Fst.landscape, landscape.to.rtheta
}


\keyword{misc}
